last big int->int16

+++ src/compat64/bios.h
@@ -7,7 +7,7 @@
 #define _KEYBRD_READY 1
 #define _KEYBRD_READ 0
 
-inline int _bios_keybrd(int cmd) {
+inline int16 _bios_keybrd(int16 cmd) {
     (void)cmd;
     return 0;
 }
