+++ src/f15.c
@@ -44,9 +44,9 @@ const char *GAME_FLIGHT = "EGAME.EXE";
 const char *GAME_DEBRIEFING = "END.EXE";
 const char *DEBUGGER = "Z:\\DEBUG.COM";
 const uint16 GFX_INIT_ARG = 2;
-const int RET_MENU = 0xc;
-const int RET_DEBRIEFING = 0x23;
-const int RET_NONZERO = -1;
+const int16 RET_MENU = 0xc;
+const int16 RET_DEBRIEFING = 0x23;
+const int16 RET_NONZERO = -1;
 enum { CMDLINE_LEN = 128 };
 char cmdlineBuf[CMDLINE_LEN] = "";
 const char FAR *CMDLINE = (const char FAR *)cmdlineBuf;
@@ -68,12 +68,12 @@ static unsigned char FAR *nullGuardPtr(void) {
 }
 static void nullGuardSave(void) {
     unsigned char FAR *p = nullGuardPtr();
-    int i;
+    int16 i;
     for (i = 0; i < NULLGUARD_SIZE; ++i) nullGuard[i] = p[i];
 }
 static void nullGuardRestore(void) {
     unsigned char FAR *p = nullGuardPtr();
-    int i;
+    int16 i;
     for (i = 0; i < NULLGUARD_SIZE; ++i) p[i] = nullGuard[i];
 }
 #ifdef NO_ASM
@@ -100,7 +100,7 @@ void game_init(void) {
     size_t freeMemory;
     int8 FAR *charPtr;
     uint16 gfxBufAddress;
-    int err;
+    int16 err;
 #ifndef NO_ASM
     OverlayFunc gfxInit = NULL;
     uint16 gfxDrvAddress;
@@ -191,8 +191,8 @@ void game_init(void) {
     LogInfo(("Initialization complete, free memory = %s", sizeString(dos_getfree())));
 }
 
-bool game_run(const char *filename, const int returnCode, const bool debug) {
-    int err;
+bool game_run(const char *filename, const int16 returnCode, const bool debug) {
+    int16 err;
     /* launch game executable */
     if (debug) {
         LogInfo(("Executing %s under the debugger", filename));
@@ -223,9 +223,9 @@ uint16 load_segment(const uint16 envParagraphs) {
     return dos_lastFreeBlock() + 1 + envParagraphs + 1;
 }
 
-int main(int argc, char *argv[]) {
+int16 main(int16 argc, char *argv[]) {
     /* process cmdline args */
-    int argIdx, charIdx;
+    int16 argIdx, charIdx;
     bool debugMenu = false, debugFlight = false, debugDebrief = false;
 
     log_set_app("f15");
