+++ src/shared/gfximpl.c
@@ -36,11 +36,11 @@ void drawLineWrapper(void) {
 /* NOTE: The original clearRect uses the dirty rect system and operates on
  * curPageSeg. We use the page index from the page descriptor to look up
  * the correct segment directly, matching gfx_switchColor's approach. */
-void clearRect(int16 *pageNum, int x1, int y1, int x2, int y2) {
+void clearRect(int16 *pageNum, int16 x1, int16 y1, int16 x2, int16 y2) {
     uint8 color;
     uint16 pageSeg;
     uint8 far *page;
-    int row, col;
+    int16 row, col;
 
     color = (uint8)pageNum[3];
     gfx_setPageN(*pageNum);
