+++ src/stalloc.h
@@ -3,6 +3,6 @@
 /* buffer alloc (stalloc.c) */
 #include "inttype.h"
 
-uint16 allocBuffer(int sz);
+uint16 allocBuffer(int16 sz);
 
 #endif /* F15_SE2_STALLOC */
